requirements.txt
boto3>=1.34.0
openai>=1.12.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.3
pytest>=8.0.0